fix(core/data-abbr): generalize implementation
