+++ tests/testFiles.json
@@ -4,16 +4,20 @@
   "spec/core/biblio-db-spec.js",
   "spec/core/biblio-spec.js",
   "spec/core/caniuse-spec.js",
+  "spec/core/data-abbr-spec.js",
   "spec/core/data-cite-spec.js",
   "spec/core/data-include-spec.js",
   "spec/core/data-tests-spec.js",
   "spec/core/data-type-spec.js",
+  "spec/core/dfn-finder-spec.js",
   "spec/core/dfn-spec.js",
   "spec/core/examples-spec.js",
+  "spec/core/exporter-spec.js",
   "spec/core/figures-spec.js",
   "spec/core/fix-headers-spec.js",
   "spec/core/github-spec.js",
   "spec/core/highlight-spec.js",
+  "spec/core/highlight-vars-spec.js",
   "spec/core/id-headers-spec.js",
   "spec/core/idl-index-spec.js",
   "spec/core/include-config-spec.js",
@@ -22,14 +26,18 @@
   "spec/core/issues-notes-spec.js",
   "spec/core/l10n-spec.js",
   "spec/core/link-to-dfn-spec.js",
-  "spec/core/linter-rules/check-punctuation.js",
+  "spec/core/linter-rules/check-charset-spec.js",
+  "spec/core/linter-rules/check-internal-slots-spec.js",
+  "spec/core/linter-rules/local-refs-exist-spec.js",
   "spec/core/linter-rules/no-headingless-sections-spec.js",
   "spec/core/linter-rules/no-http-props-spec.js",
   "spec/core/list-sorter-spec.js",
   "spec/core/location-hash-spec.js",
   "spec/core/markdown-spec.js",
   "spec/core/override-configuration-spec.js",
+  "spec/core/pluralize-spec.js",
   "spec/core/pre-process-spec.js",
+  "spec/core/pubsubhub-spec.js",
   "spec/core/ready-promise-spec.js",
   "spec/core/remove-respec-spec.js",
   "spec/core/requirements-spec.js",
